third_party/blink/web_tests/platform/generic/compositing/overflow/do-not-repaint-if-scrolling-composited-layers-expected.txt

Overflow scroll:
PASS did not repaint when unexpected
Overflow scroll with selection:
PASS did not repaint when unexpected
Marquee:
PASS did not repaint when unexpected
Overflow scroll with inline child:
PASS did not repaint when unexpected
Overflow hidden:
PASS did not repaint when unexpected
